feat(wasm): extract lg-host.js + host.html from wasm.go template
Pulls the inline <script> body and HTML template body out of wasm.go
into standalone pkg/rt/wasm/lg-host.js and pkg/rt/wasm/host.html,
embedded via //go:embed and inlined back into the template at build
time. Output index.html is byte-identical to the prior output
(verified by golden file test, excluding the per-build WASM blob
which carries non-determinism from the toolchain).

Step 1 of #79. The LetGoHost registration API ships in a follow-up
commit; this one is pure source-layout refactor.

Mechanics:
- New package pkg/rt/wasm exposes LgHostJS and HostHTML strings via
  //go:embed.
- The HTML template references __LG_HOST_JS_BODY_PLACEHOLDER__,
  __WASM_EXEC_JS__, and __WASM_GZ_B64__ markers that AssembleHTML
  substitutes via strings.Replace.
- TestAssembleHTMLGolden guards byte-identicality of assembled output
  against a fixed-input golden file (regenerate with
  go test ./pkg/rt/wasm -update).
- TestMarkersGone is a belt-and-suspenders check that all placeholder
  markers are fully substituted out.
